requirements.txt
torch
transformers
diffusers
gradio
python-dotenv
